rename core_atomistic and core_atomistic_qt separate tests for core
