+++ .gitignore
@@ -1,17 +1,7 @@
 # Default ignored files
 .idea/
 src/__pycache__/
-doc-code/build
-doc/gui4dft.aux
-doc/gui4dft.bbl
-doc/gui4dft.blg
-doc/gui4dft.log
-doc/gui4dft.synctex.gz
-doc/gui4dft.toc
-doc/gui4dft.out
-executable/dist/
-executable/build/
-executable/gui4dft.spec
 tests/.coverage
+tests_core/.coverage
 .coverage
 coverage.xml
