+++ src/core_atomistic/project_file.py
@@ -1,6 +1,6 @@
 import os
 import numpy as np
-from core_gui_atomistic.atomic_model import AtomicModel
+from core_atomistic.atomic_model import AtomicModel
 
 
 class ProjectFile(object):
