+++ src/core_atomistic_qt/pyqtgraphwidget.py
