+++ src/core_atomistic_qt/pyqtgraphwidgetimage.py
