+++ src/src_critplot/models/atomic_model_cp.py
@@ -7,9 +7,9 @@ import numpy as np
 from copy import deepcopy
 from numpy.linalg import inv
 from src_critplot.models.critical_point import CriticalPoint
-from core_gui_atomistic.atom import Atom
-from core_gui_atomistic.atomic_model import AtomicModel
-from core_gui_atomistic import helpers
+from core_atomistic.atom import Atom
+from core_atomistic.atomic_model import AtomicModel
+from core_atomistic import helpers
 
 
 class AtomicModelCP(AtomicModel):
