+++ src/src_critplot/qtbased/guiopengl_cp.py
@@ -5,8 +5,8 @@ from typing import Callable
 import OpenGL.GL as gl
 import OpenGL.GLU as glu
 import numpy as np
-from core_gui_atomistic_qt.guiopenglbase import GuiOpenGLBase
-from core_gui_atomistic.helpers import is_number
+from core_atomistic_qt.guiopenglbase import GuiOpenGLBase
+from core_atomistic.helpers import is_number
 from src_critplot.models.atomic_model_cp import AtomicModelCP
