--- tests/core_gui_atomistic/conftest.py
@@ -1,8 +1,0 @@
-from pathlib import Path
-from core_gui_atomistic.periodic_table import TPeriodTable
-import pytest
-
-
-@pytest.fixture
-def period_table() -> TPeriodTable:
-    return TPeriodTable()
