+++ tests_core/core_atomistic/conftest.py
@@ -1,0 +1,21 @@
+from core_atomistic.atomic_model import AtomicModel
+import pytest
+
+
+@pytest.fixture
+def h2o_model() -> AtomicModel:
+    #         x    y    z   let charge
+    atoms = [[0.0, 0.0, 0.0, "O", 8], [-1.0, 0.0, 0.0, "H", 1], [1.0, 0.0, 0.0, "H", 1]]
+    return AtomicModel(atoms)
+
+
+@pytest.fixture
+def swnt_33() -> AtomicModel:
+    #         x    y    z   let charge
+    atoms = [[-0.70056811, 1.92479506, - 0.61920816, "C", 6], [-1.31663736, - 1.56910731, - 0.61920816, "C", 6],
+             [-2.01720547, - 0.35568775, - 0.61920816, "C", 6], [2.01720547, - 0.35568775, - 0.61920816, "C", 6],
+             [1.31663736, - 1.56910731, - 0.61920816, "C", 6], [0.70056811, 1.92479506, - 0.61920816, "C", 6],
+             [-1.31663736, 1.56910731, 0.61920816, "C", 6], [-0.70056811, - 1.92479506, 0.61920816, "C", 6],
+             [-2.01720547, 0.35568775, 0.61920816, "C", 6], [2.01720547, 0.35568775, 0.61920816, "C", 6],
+             [0.70056811, - 1.92479506, 0.61920816, "C", 6], [1.31663736, 1.56910731, 0.61920816, "C", 6]]
+    return AtomicModel(atoms)
