+++ tests_core/core_atomistic/test_periodic_table.py
