+++ tests_core/core_atomistic_qt/test_pygrimagewidget.py
